fix(calendar): move sync cron to a node-only module to unbreak edge bundle
Putting the heavy import chain (calendar-sync → node-ical → fs / dns/promises,
+ redis → dns/promises, + crypto → node:crypto) directly in instrumentation.ts
caused webpack to try to bundle those node-only modules into the EDGE runtime
bundle, which has no Node built-ins. Build failed with
  Module not found: Can't resolve 'dns/promises'
  Module not found: Can't resolve 'fs'
  Reading from "node:crypto" is not handled

Move the cron into src/lib/server/calendarSyncCron.ts. instrumentation.ts
dynamically imports it only inside the NEXT_RUNTIME === 'nodejs' branch, so
webpack dead-code-eliminates that import from the edge bundle entirely.
